+++ packages/extension-koni-ui/src/Popup/Home/Earning/EarningPositionDetail/index.tsx
@@ -216,8 +216,12 @@ function Component ({ compound,
       return true;
     }
 
-    return false;
-  }, [poolInfo.chain, poolInfo.type]);
+    return !poolInfo.metadata.availableMethod.join;
+  }, [poolInfo.chain, poolInfo.type, poolInfo.metadata.availableMethod.join]);
+
+  const isDisabledUnStaked = useMemo(() => {
+    return !poolInfo.metadata.availableMethod.defaultUnstake && !poolInfo.metadata.availableMethod.fastUnstake;
+  }, [poolInfo.metadata.availableMethod.defaultUnstake, poolInfo.metadata.availableMethod.fastUnstake]);
 
   const subHeaderButtons: ButtonProps[] = useMemo(() => {
     return [
@@ -284,6 +288,7 @@ function Component ({ compound,
         <div className={'__transaction-buttons'}>
           <Button
             block={true}
+            disabled={isDisabledUnStaked}
             icon={(
               <Icon
                 phosphorIcon={MinusCircle}